Various changes to cpan/ExtUtils-MakeMaker
Fix code which modifies a hash while each()ing over it

It has never been correct to modify a hash during traversal by each().

In older perls this was unlikely to break the tests. In newer perls
it is relatively likely, and in 5.18 it will be almost certain.

Do not hard code the list of perl header files - discover them from disk instead

Don't need completely different include file collector on VMS.

We just need to skipping putting a Unix delimiter between the
directory and the filename.

Includes a version bump to all the EU::MM files (to prevent any possible
$VERSION consistency checks from failing like they would on *nix)

This patch has been submitted to the EU::MM project in a modified form
and once merged there should replace these changes.