semi-Randomize hash insertion order
Use HvRAND(hv) as a set of flag bits. Each time we use a bit we
rotate the value of HvRAND(hv) so the next time we use a different
bit. This also means that inserting into a hash has a high chance
of changing the perceived iteration order of the hash.