Revert "Define RXf_SPLIT and RXf_SKIPWHITE as 0"
This effectively reverts commit cccd1425414e6518c1fc8b7bcaccfb119320c513.

We dont want to get rid of these vars, and it turns out that
RXf_SEEN_LOOKBEHIND is used only in the same situation of the new
RXf_MODIFIES_VARS. So I have renamed RXf_SEEN_LOOKBEHIND to
RXf_NO_INPLACE_SUBST, and then instead of using two vars we use
only the one. Which in turn allows RXf_SPLIT and RXf_SKIPWHITE to
have their bits back.

This is in preparation for reverting another patch in this sequence.