Revert "[perl #94490] const fold should not trigger special split " ""
Revert, with minor changes and some merges: 5255171e6cd0accee6f76ea2980e32b3b5b8e171

The reverted patch was meant to make:

        split( 0 || " ", $thing )            #1

consistent with

        my $x=0; split( $x || " ", $thing )  #2

and not with

        split( " ", $thing )                 #3

This was reverted because it broke C<split("\x{20}", $thing)>, and
because one might argue that is not that #1 does the wrong thing,
but rather that the behavior of #2 that is wrong. In other words
we might expect that all three should behave the same as #3, and
that instead of "fixing" the behavior of #1 to be like #2, we should
really fix the behavior of #2 to behave like #3.

Also, it doesn't make sense to move this further from the regex
engine. We really want the regex engine to decide this stuff itself,
otherwise split " ", ... wouldn't work properly with an alternate
engine. (Imagine we add a special regexp meta pattern that behaves
the same as " " does in a split /.../. For instance we might make
split /(*SPLITWHITE)/ trigger the same behavior as split " ".

Anyway, it can wait until after 5.18 to figure it out. The op logic
involved in this stuff makes my head spin.